Reafactor Business API Calls
